Add hooks, logic, and styling to listing card
The reservation start and end dates are stored
and rerendered with memo hooks. The onAction
function prop is called within a use callback
hook. An image component with the desired
styling and animation is included in the listing
card component to show the listing's image found
from imageSrc (stored in cloudinary).