Create useFavorite hook and favorites route.
The route is use to add and delete favorites from
the database, where the logic for clicks in
the heart button component are handled in the custom
useFavorite hook. If the user isnt logged in, the
login modal is opened with the login modal hook.
Within the route, post and delete functions
add and delete the favorites from the current
users favorites list. An error related to the need
to await the params prop in the route was fixed.