Add disabled date function in Listing Client component
This function returns an array of the dates that
aren't available for new reservations. It
uses the reservations array prop to find all the
current reservations, and with each reservation
object comes a start and end date. The eachDayOfInterval
finds all the date objects between those dates,
and adds those dates to the resulting dates array.