Add properties page w/ PropertiesClient
The page uses the getListings action to retrieve
the current user's properties from the db. The
listings and user are passed as props to the client.
The client's onCancel function uses an api w/ axios
to delete the listing from the db. The listings are
shown with Listing Card components.